getty/telnet/tftpd: minor cleanup.
Remove STDIN_FILENO and use FLAG().

Use xsetspeed() rather than have a duplicate table.

Inline getty's print_prompt().

There's still a lot of cleanup needed here, in particular use of toybuf
and removal of the inappropriately-named HOSTNAME_SIZE (and ideally
sharing that logic with login(1) which already contains a better
implementation of it).