cpio: Don't lchown() if -t is specified
When using -t to inspect an archive, cpio would try to set the owner of
any symlink in the archive, even though the symlink wasn't created by
the command previously.

This would lead to two results, either the command fails with a "No such
file or directory" message when trying to lchown() the symlink path, or
an existing file, with the name of the symlink, is lchown()-ed.

Guard the lchown() function call with a "if (!FLAG(t))" block, and add
regression test for this.