syslogd: Ensure that socket sends are '\n'-terminated
The documentation for syslog states that "[t]he format string need not
include a terminating newline character", which implies that syslogd is
responsible for ensuring one. Busybox ensures this terminating newline
always, so add code to ensure it for the one case (sockets) where we
didn't.

Also, trim any trailing '\n' and '\0' characters from the read buffer,
to avoid (sometimes) printing multiple newlines.