rmmod: various fixes.
My original intent was to just get rid of the meaningless O_EXCL, which
isn't used by the kernel, after noticing that it's passed here but not
in modprobe's module removal code.

But then I noticed that other rmmod implementations allow you to pass a
list of modules, not just one (whereas we were ignoring all but the
first). And that the help syntax was wrong. And that strend() lets us
write the code to remove the ".ko" suffix a bit nicer.

And of course, switch to the FLAG() macro.