insmod: tiny cleanup.
finit_module(2) was added in Linux 3.8, and even CentOS 7.2 has 3.10, so
we should be able to assume that now rather than working around its
possible absence (especially in the *build* environment). I've left the
runtime workaround for now, since we need that to support the stdin use
case anyway.

Also add a missing error check for a failure to read, so we can give a
more relevant error message (we can't use xreadfile() because we need to
know the resulting size).