strace: initial commit.
This is most notably lacking -p and -f, and only covers a small subset
of system calls, but it's "useful" in the sense that you can strace
something like toybox date, say, and see everything.

There's a fundamental assumption here that we don't need to worry about
multiple personalities. (This is what makes "real" strace's build so
complex, and why Android has _never_ actually shipped a fully correct
strace build.) Given that the desktop's already been 64-bit only
for a while, microcontrollers are likely to stay 32-bit only for a
while, and the odd one out -- mobile -- is moving to be 64-bit only
(https://www.arm.com/blogs/blueprint/64-bit), so by the time this is
ready to graduate from pending, I'm assuming no-one will care!