strace: fix macOS build.
Obviously we're never going to be able to build strace for macOS,
but right now the reference to the struct was breaking even a macOS
build with strace disabled! From poking at building for arm on the
Raspberry Pi over the weekend, changes like this will be needed
there too (because different architectures have different _names_
for this struct, not just different contents).

Tested on macOS (where it juts fixes the build) and on x86-64 (where
strace continues to function as before).