blkid: more scanf fun.
The unbounded %[] here made me look at the kernel, where the size of a
block device name is limited to 32 bytes (including the NUL), so
although the 1024 bytes of toybuf was plenty, that meant that the
32-byte buffer for "/dev/%s" was too small because the "%s" alone could
require the entire buffer. Fix that, and simplify the code a bit by only
using toybuf for the fgets(). I didn't switch to fscanf() because the
file has two lines of headers, and calling fgets() twice to skip the
header seemed more awkward than keeping the continue in the loop.