tftpd: fix the parsing of options
TFTP supports passing options such as 'blksize' as part of the initial
RRQ or WRQ packet. These options consist of a null-terminated name
followed by a null-terminated value. The current code segfaults due to
mis-calculating the address of the value.

Add a `next_token` helper that advances past the next null and also
detects advancing past the end of the packet.

Tested by running:

    atftp -g -r /toybox -l toybox --option "blksize 1500" --option "foo bar" localhost 8069

Signed-off-by: Michael Hope <mlhx@google.com>