gpiod: new commands.
Equivalent to the content of the gpiod debian package. Very different
interface to the traditional Raspberry Pi command, but that command
explicitly calls itself deprecated and tells you to use the gpiod
package's commands instead.

"Works for me" on a Raspberry Pi 400, with a (small) variety of things
connected to GPIOs as inputs and outputs.

Missing gpiomon, which -- since the rest of these are about a year old
at this point -- it doesn't look like I'm going to get around to writing
any time soon unless someone asks for it...