file: don't be confused by BSS sections.
BSS sections are *expected* to claim a larger size than is actually
present in the file. Unlike program headers which have two different
size fields for the memory and file sizes, sections headers only have
one size field, so we're using the right field; we just need to ignore
it for our overflow checking purposes.