file(1): call out apparently invalid ELF files.
Realistically, corrupt ELF files are rare enough that it's more likely
that we've actually found a bug in file(1), such as the .bss bug fixed
in the previous patch. That bug went undiscovered so long because we
silently give up on ELF files we don't understand, and the output didn't
look obviously broken even if you'd been told it was broken and were
looking at it.

Give up noisily instead so we find and fix any future bugs more quickly.

Also remove some duplicated code: the commit that switched to mmap(2)
rather than lseek(2) had two copies of the mmap call.