Apparently macOS-12.4 isn't an option.
But it's not an _error_ --- github just waits forever for a runner
to be available? Anyway, from
https://github.blog/changelog/2022-04-25-github-actions-public-beta-of-macos-12-for-github-hosted-runners-is-now-available/
it looks like `macos-12` should work.

There's also `macos-latest`, but it sounds like that's still
`macos-11` atm, so although we should probably come back and switch
to that, it's a bit early yet.

Although we could switch to `macos-latest` but explicitly say that
we need the homebrew coreutils installed?