Very small MVP of git clone
I have a small MVP. It clones version 0.0.1 of toybox and writes it to the
working directory. I dont get it running for later versions. Most likely
due to a (late occuring) bug (base object not found) and/or a heap overflow
in resolving_deltas(). Also version 0.0.1 finishes, but does not exit.
Could that be a some crash in the clean up of corrupted memory when toybox
exits?

I sent you this version as it show the git clone MVP for version 0.0.1 of
Toybox.  The delta resolving debugging is really a point in the as, because
I dont have a fixpoint for testing. It is crashing for later version,
because it does not find the base object for a delta object to be resolved.
Initially this error was cause because my libc bsearch() extension inserted
wrong after some thousand objects. I created my own bsearchpos() for
insertation. In addition the base object lookup can fail, when the base
object itself is a delta object to be resolved and the subsequent resolving
went wrong and ended in a different hash.

I observe that even version 0.0.1 files contain some more bytes at the end
than they should have..anyways version 0.0.1 can lookup all the
recalculated hashes and recreate the files. It only messes up for big
chains of delta resolving. There I dont have a fixpoint for testing and no
good test example, because changing on byte might change all the hashes and
git itself is not very verbose when resolving delta (e.g. as a reference
for testing).

I hope that I do a obvious messup of memory there. I tried to identify it
with valgrind, but it just complains some uninitialized variable and
unfreed memory. I struggled to activate GCC sanatizing in make.sh to get
indication for the location of the error.

I think if is time to let you sneak peek into the current version. I did
additional commenting and linked docu to make it understandable. I am prood
that version 0.0.1 is checked out for the current general implementation,
but I am also unhappy how broken it feels for everything else.

Could you have a look at it and could

a) decide if it is "good enough" for /pending or
b) could you give me indication how to set sanatizing cflags for make.sh. I
tried by setting $CFLAGS and make seem to recognize the change, but it seem
the binary does not contain the sanatizing support (is it only in the
unstripped version)?