Avoid copy_file_range(2) on Android.
Strictly we could use it if the target API level is high enough,
but that would mean that you couldn't safely run the resulting
binary on an older Android release, and anecdotally I don't think
people are that careful.

The obvious alternative would be to install a signal handler for
SIGSYS so we can probe for copy_file_range(2) at runtime rather
than compile time, but it's not obvious to me that it matters enough?