Enlarge macos_defconfig.
Add everything that uses xgetrandom(), and also everything that has
tests (but isn't still in pending) to the macos_defconfig.

I've sorted the list because it's impossible to manage otherwise.

All of the tests pass, except for the tar tests that were already
failing because there's no *group* called "root" on macOS (group 0
is called "wheel" instead). I haven't fixed that because it's quite
invasive, and TIL we're already ignoring macOS test failures in the
github CI (which is why we hadn't noticed this). There's also one
failure in the truncate tests: macOS doesn't do sparse file extension.