Fully commented version of very small git clone.
It has 1 or more bugs (most likely memory-related) in the delta-resolving.
I am sure that it is memory-related and in the resolving , because I
concatenate deltas and the result is slightly broken, which creates wrong
hashes and confused the resolving further. Repos & versions without deltas
work beautifully (i.e. toybox 0.0.1 which is the only version of toybox
without deltas).

Due to 3 kinds of indirections (hashes-references, deltas and recursive
resolving) it is hard to test and to debug. In addition there are not
reference examples to write generic test cases for delta resolving.