Fix grep to not truncate last character of file provided by -f argument
The command line argument -f allows a user to specify a file that lists
multiple regular expressions (one per line) to match. It turned out
that, if the file ended with a new line character (like most text files
do), then the character just prior to that newline character got
truncated.

For example, if the file was just 4 bytes long and contained "abc\n",
then the character "c" would be removed.