ls -N.
Not everyone wants -b, but that's the default for a tty, and we don't
currently have an opt out.

To my surprise, -N (despite having the long name --literal) only
disables -b, not -q. There's a separate --show-control-chars (opposite
of --hide-control-chars, which is the long name for -q) for that, so -N
really is just "the opposite of -b".

Since no-one's asked for --show-control-chars yet to my knowledge, I've
not implemented that here.

Also add tests for -b/-q/-N.