modprobe: Make /proc/modules optional
Currently, /proc/modules is a strong requirement for loading modules.
This is problematic as procfs might not always be available. Checking
/proc/modules only allows to flag if a module is already present to
avoid loading it. It's something the kernel handles well already
(finit_module returns EEXIST). So if this might be good to have, not
being able to check the module presence shouldn't prevent modprobe from
trying.