Add bash -c wrapper around run-qemu.sh call.
Timeout doesn't know how to exec a shell script that doesn't start
with #!/bin/sh but I don't want to add that line to the script generation
because I'd have to explicitly specify bash/sh/mksh.

On systems with the Defective Annoying SHell the /bin/sh symlink is simply
broken and should not be used, but mksh sort of works and when your current
shell is bash running a script from the command line will use bash even
when it's not /bin/bash but /strange/freebsd/path/to/bash instead (and
no, "env is at a known path but bash isn't" still doesn't make sense).

NOT specifying which shell to use fixes more than it breaks, so far...