git.c progress
I "solved" the issue, which made the git.c code not finding hashes and
returned NULL (I used a modified/buggy bsearch() instead of the indented
bsearchpos()). As a result it now clones Toybox 0.0.1 and 0.0.2.
bsearchpos() never returns NULL, but a long of the position nearest to the
hash. The drawback is that it can return wrong hashes, if it does not find
the intended hash, but the good thing is that I can debug the wrong hashes
in much better and interpretable way in git checkout, if the code does not
crash before git checkout.

When cloning toybox 0.0.2, it shows unintended random bytes at the end of
the recreated source files, which make the files too long. It indicated
that the delta resolving is not correct yet and will result in wrong
hashes. It is also confirmed by higher version than 0.0.3.

For version 0.0.3 and higher the git fetch "seems" to work "perfectly", but
when checking out, some files are interpreted folders. I guess this comes
from the files with unintended random bytes in the end, which result in
wrong hashes, which bsearchpos() resolves the next similar hash. The good
thing is that I know at this level which expected file/hash it was and how
it should be resolved. Much better starting point to debug delta resolving
than if you have to do this in fetch while delta resolving. The code feels
also much more complete seeing how parts of the repository are already
checked out.