init: Remove trailing newline when parsing inittab
Commit 05e4f52c3c modified init to use the standard getline instead of
get_line when parsing the inittab. However, these functions have
different behavior that wasn't properly accounted for: get_line strips
the trailing newline, and getline keeps it (if it exists, i.e. not EOF).

This led to a bug where every command in the inittab is parsed as ending
with "\n". This commit corrects init to remove the newline, fixing this
issue.