login: Fix segfault if unknown username is provided
The 'login' command segfaults if the user provides an invalid username.
This happens because 'getpwnam' returns NULL, and this is stored in a
local variable. Later, after the password check fails, the pointer is
dereferenced to produce the syslog message.

This commit fixes the bug by instead using the 'username' variable in
the log message, which is never NULL.