login: Don't prompt for password if shadow password is empty
The 'login' command doesn't prompt for a password if the password is
empty. However, the current implementation only checks if the
/etc/passwd entry is empty. If the /etc/shadow entry is empty, it still
requires the user to enter an empty password.

This commit addresses the issue, and no longer prompts for a password if
the /etc/shadow entry has an empty password.