vi: don't call open() with O_CREAT but no mode.
This was invisible before because we were calling xopen(), which
always passes mode 0 to open(). Now we're calling open() directly,
bionic's _FORTIFY_SOURCE spots this.

Rather than create with 0600 and then chmod() later as the code
used to do, move the stat() before the creation so we can supply
the final permissions from the beginning.

Also rewrite the default permissions in octal, in keeping with
toybox style.