ps: widen the VSIZE field for Android's default view.
On LP64 systems, address space is effectively free, and the security
folks have found a variety of ways to use large PROT_NONE regions. For
example, Android currently spends 3GiB on the CFI shadow and 7GiB
padding the scudo heap, meaning that even toybox cat has a VmSize around
10GiB. This means that the default ps output doesn't line up.

Arguably this should be a change to the default width (currently 7) of
the VSZ entry in the "typos" array, but as long as desktop Linux isn't
using its address space for security mitigations, that seemed like it
might be too contentious.