Change FLAG(x) macros to always return 0 or 1. Treewide audit of FLAG() users to make sure nobody NEEDED FLAG(x) to return the masked bit value, and to remove now-redundant !!, with a number of in-passing cleanups while I was there.
I think I fixed an actual bug in patch.c (-R depended on the flag value
but commit 6f6b7614e463 changed it, so "+-"[FLAG(R)] was comparing
against the NUL terminator when measuring fuzz) but we didn't have a
test for "fuzz autodetection". And there was another bug in tar.c
where DIRTREE_BREADTH was assigned to a local variable... and then not
used. But the test passed? (Not sure the flag's needed at the top level,
added a comment to the test suite to remind me to revisit that.)

While I was there, capitalize TODO comments so they're easier to grep for.