Fix error reporting for find -commands that take an argument but haven't got one because they're at the end of the command line.
This unknown commands, because the logic is an if/else staircase handling
commands that reaches an "else { everything after here takes an argument"
that pops the argument, then has an eventual else goto error; for unknown
commands. Since argv[] always ends with a NULL pointer, we were trying
to print that in the error message.