gpiod: fix foreach_chip crashes.
This was segfaulting because the loop copying the chips into the array
was corrupting the list, and because the llist_traverse was being given
the wrong pointer.

The strdup wasn't failing, but should be an xstrdup anyway. (I think
all the strdups in toys/ should actually be xstrdups. Maybe we should
even poison strdup, or just `#define strdup xstrdup`?)