Reorganize patch logic so EOF detection works with fuzz support.
Comparison involves buffering lines we're unsure of, and when a match
fails several lines in, patch has to write out the first line and recompare
the hunk to the remaining buffered lines because the actual start of this
hunk may have been fuzzed past earlier. EOF detection was still evaluated
immediately instead of as part of that deferred re-examining (it made more
sense before fuzz support was added), and the "we're not at the end of the
hunk" test was happening too early.

While I was there, replace the -x debug code with -vv so you can ask for
maximum verbosity if you want to see patch's decision making process.
(Hopefully somewhat intelligible for normal users now?)