Add some (void) typecasts to shut up busybody compiler "unused result" warnings.
If watchdog writes fail, the system is about to reboot. No recovery possible.

If the init task can't dup() stdin to create stdout and stderr (when
we just closed the old ones, if any), your kernel has bigger problems.

The fgets() in netstat are reading /proc data produced directly by the kernel.
Each is discarding a line of input from /proc (because the kernel devs
decided to put a pointless header line on the data to make it harder to
machine parse the output, breaking multiple rules from half of The Art
of Unix Programming), and in each case it's immediately followed by a
loop reading the rest of the lines of input, which DOES respond to each
result. Not only do we know the kernel is generating this for us,
but if the first read somehow returns error but the second read produces
good data... what did that error mean exactly?

We _explicitly_ do not care about these return results. There's no useful
"error handling" we could put in any of these places, so SHUT UP about it.