Fix pidof tests when stdout is a socket
The pidof tests try to block a subprocess forever by reading from
stdout via /proc/self/fd/1.  This works for most stdout fds, but
fails if stdout is a socket, which happens when running the tests
in a sandbox in the Android build.  Create a named fifo and read
on that instead.