microcom: simple menu, new "paste file" functionality.
Rather than take up yet another key for the new functionality, I've moved
everything to a menu, similar to the one in telnet. It's not amazing,
but it's the least code I could get away with, and it's good enough for
now, and it leaves us in a better place if we need to add more stuff
(such as xmodem, say).

But this is enough to let me stop using minicom and get back to toybox
microcom for today!