strace: fix arm64 test.
We're not specifically looking for armv8; we want any aarch64.

I've also changed the 32-bit arm test because (a) Linux doesn't
support arm32 OABI any more, (b) it's clearer this way that 32-bit
armv8 still belongs in this branch, and (c) __arm__ and __aarch64__
are the usual spellings everywhere. (Sure, the latter is ugly, but
we have __i386__ rather than __x86__ one line later anyway! I still
have to grep AOSP for mistaken uses of the non-existent __x86__ on
a regular basis. And don't get me started on __riscv and __riscv_xlen!)