Add second stage toolchain to command line dependencies.
Because we statically link the binaries against musl-libc, we need a musl-libc
toolchain for the host. (Static linking is broken on modern glibc.)

The -host toolchain we build first has various shortcuts to remove
incestuous dependencies, and the result is mostly good for building the
full toolchain. (Because gcc is rickety and overcomplicated. Yes there's
a pattern here: "If it starts with g, scream and flee...")

Thus non-host toolchains need the full host toolchain to (reliably) build,
and if you did a fresh clone and built a specific toolchain ala
"~/toybox/scripts/mcm-buildall.sh or1k::" it needs both host toolchains,
so add it as a dependency when you specify toolchains to build on the command
line. (In theory, if the toolchain exists it shouldn't rebuild it. You have
to rm -rf the corresponding dir/sqf out of ccc to force a rebuild.)