Various help fixes.
Ensure that the usage lines include everything mentioned below in
the detailed help, and vice versa.

Plus a couple of clarifications (like what's the offset for
fallocate, and the fact that the length is mandatory).

For the ones I've touched, I've improved the consistency of
alphabetization, but I haven't alphabetized the ones I had no other
reason to touch.