Move the modules archive under docs.
If you build kernel modules, it archives them up separately and then appends
them to the initramfs.cpio.gz so you don't have to rebuild the kernel
to rebuild the filesystem. But this leaves a file lying around at the top
level which is not used to launch qemu, and those go in "docs".